+++ CHANGELOG.md
@@ -14,6 +14,9 @@ This version of the dashboard requires Cumulus API >= feature/duplicate-granule-
   - Populated producer ID in dash/db fixtures
   - Updated the dashboard to use alpha version `@cumulus/api@20.1.3-alpha.2` and
     `@cumulus/aws-client@20.1.3-alpha.2`
+- **CUMULUS-4086**
+  - Added fixture and test case for cypress tests to test if duplicate producerGranuleIds will show in the
+    granules' producerGranuleId column
 
 ## [Unreleased]